### 2025-09-30 #9110 @editorialbot
[PRE REVIEW]: Plaquette: An Object-Oriented Framework for Embedded Signal Processing in Interactive Media
**Submitting author:** @sofian (<a href="http://orcid.org/0000-0001-6621-6427">Sofian Audry</a>)
**Repository:** https://github.com/SofaPirate/Plaquette/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.8.0
