### 2025-10-01 b9d6729
Updated .bib file with DOI and adjusted error in year.
