+++ extras/joss-paper/paper.bib
@@ -1,27 +1,33 @@
 @inproceedings{Audry2020-Behaviour,
-  ids = {Audry2020-Behavioura,Audry2020-Behaviourb},
+  ids = {Audry2020-Behaviour},
   title = {Behaviour {{Aesthetics}} of {{Reinforcement Learning}} in a {{Robotic Art Installation}}},
   booktitle = {4th {{NeurIPS Workshop}} on {{Machine Learning}} for {{Creativity}} and {{Design}}},
   author = {Audry, Sofian and Dumont-Gagné, Rosalie and Scurto, Hugo},
   date = {2020-12},
   location = {Vancouver, Canada},
   url = {https://hal.archives-ouvertes.fr/hal-03100907},
-  urldate = {2021-01-29}
+  urldate = {2021-01-29},
+  hal_id = {hal-03100907},
+  hal_version = {v1}
 }
 
-@incollection{Audry2024-Choreomata,
+@inbook{Audry2023-Choreomata,
+  ids = {Audry2024-Choreomata},
   title = {Choreomata},
-  booktitle = {Choreomata: Performance and Performativity after {{AI}}},
+  booktitle = {Choreomata},
   author = {Audry, Sofian},
-  editor = {Trillo, Roberto Alonso and Poliks, Marek},
-  date = {2024},
-  edition = {First edition},
+  date = {2023-12-04},
+  edition = {1},
   pages = {283--307},
-  publisher = {CRC Press, Taylor \& Francis Group},
+  publisher = {{Chapman and Hall/CRC}},
   location = {Boca Raton},
+  doi = {10.1201/9781003312338-16},
+  url = {https://www.taylorfrancis.com/books/9781003312338/chapters/10.1201/9781003312338-16},
+  urldate = {2025-10-01},
+  bookauthor = {Trillo, Roberto Alonso and Poliks, Marek},
   isbn = {978-1-003-31233-8},
   langid = {english},
-  annotation = {OCLC: 1415574296}
+  pagetotal = {25}
 }
 
 @book{Banzi2022-Getting,
@@ -54,20 +60,24 @@
 }
 
 @article{Candy2018-PracticeBased,
+  ids = {Candy2018-Practicebased},
   title = {Practice-{{Based Research}} in the {{Creative Arts}}: {{Foundations}} and {{Futures}} from the {{Front Line}}},
   shorttitle = {Practice-{{Based Research}} in the {{Creative Arts}}},
   author = {Candy, Linda and Edmonds, Ernest},
-  date = {2018},
+  date = {2018-02},
   journaltitle = {Leonardo},
+  shortjournal = {Leonardo},
   volume = {51},
   number = {1},
-  eprint = {26808545},
-  eprinttype = {jstor},
   pages = {63--69},
   publisher = {[Leonardo, The MIT Press]},
-  issn = {0024-094X},
-  url = {https://www.jstor.org/stable/26808545},
-  urldate = {2025-09-29}
+  issn = {0024-094X, 1530-9282},
+  doi = {10.1162/LEON_a_01471},
+  url = {https://direct.mit.edu/leon/article/51/1/63-69/46472},
+  urldate = {2025-10-01},
+  abstract = {This article explores the subject of practice-based research, its application in the creative arts and its role in generating new forms of knowledge in the context of the PhD. Our aim is to provide more clarity about the nature of practice-based research, the approach we advocate and how it contributes to new knowledge that can be shared and scrutinized in a form that is both accessible and rich in its representation of the full scope of creative arts research. We draw on examples spanning over 35 years of experience in supervising interdisciplinary PhD research programs in the arts, design and digital media.},
+  issue = {1},
+  langid = {english}
 }
 
 @inproceedings{Gee2023-BioSynth,
