### 2025-10-01 #9110 @sofian
@danielskatz Thank you. I compiled it locally using docker but I think I misused the tool, I'll fix the problems and use the bot.
