### 2025-10-01 #9110 @sofian
I may suggest the following potential external reviewers (listed in alphabetical order):

Leah Buechley at University of New Mexico https://www.cs.unm.edu/directory/faculty-profiles/leah-buechley.html
Petra Gemeinboeck at Swimburne University of Technology https://experts.swinburne.edu.au/6197-petra-gemeinboeck
Kate Hartman at University of Toronto https://www.ocadu.ca/academics/explore-faculty/khartman
Joseph Malloch at Dalhousie University https://www.dal.ca/faculty/computerscience/faculty-staff/joseph-malloch.html
James Partaik at Univesity of Quebec in Chicoutimi https://www.linkedin.com/in/james-partaik-3a556b43/
