### 2025-10-01 #9110 @danielskatz
@sofian - a couple of minor comments as I read your paper

* "softwares" is not English - this could be "software packages" or just "software" (which can be singular or plural)
* you need to add a "References" heading at the end of your .md file so that this is included when the PDF is built

Otherwise, this seems ready to move to an editor, and having reviewer suggestions is useful. Given our current editor load, I think I will edit this myself.
