### 2025-10-01 8c0066d
Fixed typo and added references heading at the end.
