### 2025-10-04 #9110 @danielskatz
👋  @LaurenceDenneulin & @ixjlyons & @timonmerk, would one or more of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
