### 2025-10-21 SofaPirate/Plaquette#152 @timonmerk
[JOSS Review] documentation
This is part of the review at JOSS https://github.com/openjournals/joss-reviews/issues/9144#issuecomment-3423665667

First of all, amazing project! I installed the software on an Arduino, tried out some examples, and they all worked, were well documented, and extended with an intuitive syntax the Arduino core libraries.

I think my main suggestion would be to add within the paper a comparison to different existing Arduino libraries.
After a brief search I came across the following filtering libaries:
 - https://github.com/AsheeshR/Microsmooth
 - https://github.com/kosme/arduinoFFT
 - https://github.com/sebnil/FIR-filter-Arduino-Library
 - https://github.com/arc12/LibSimpleFilters
 - https://github.com/RobTillaart/functionGenerator
 - https://github.com/jeroendoggen/Arduino-signal-filtering-library

and existing timing libraries:
 - https://github.com/AlexandreHiroyuki/DataTome
 - https://github.com/brunocalou/Timer

I know a lot of them are currently not well maintained, or do not address some functionality of Plaquette, but I still think it would be great to reference some of them, their short-comings, and what Plaquette adds in addition.

I would add also a comparison to the existing Arduino/ATMega32 timer Interrupt system:
 - https://deepbluembedded.com/arduino-timers/
How is plaquette simplifying interrupt understanding and their applications? I think that is definitely a strength of the library and simplifies timing applications!

Next some minor comments. It would be great to add a Contributing Section to the paper and reference to the GitHub issue tracker.
The [JOSS review](https://joss.readthedocs.io/en/latest/review_criteria.html) requires such a section within the documentation:

> Community guidelines[](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)
> There should be clear guidelines for third-parties wishing to:
>
> Contribute to the software
>
> Report issues or problems with the software
>
> Seek support

Also some documentation questions. For the `wrap` function, the documentation specifies the examples https://plaquette.org/wrap.html?highlight=wrap `wrap(13.0, 1.0, 5.0) ->  3.0` and `wrap(10.0, 1.0, 5.0) -> 1.0` Maybe I'm missing something, but wouldn't the values be clipped within the boundaries? So (13.0, 1.0, 5.0) ->  5.0?

Also I found in https://plaquette.org/timing_basics.html a typo `proprties`
