### 2025-10-22 b2659c0
Added draft code of conduct.
