### 2025-10-22 428e038
Added draft instructions for contributing (besides coding standards).
