### 2025-10-22 83af5c7
Added GitHub issue templates.
