+++ .github/ISSUE_TEMPLATE/quick_note.yml
@@ -1,0 +1,15 @@
+name: "✏️ Quick note"
+description: Leave a quick thought, glitch, idea, or question — one box only
+title: "[Note] <short summary>"
+labels: ["triage"]
+assignees: []
+body:
+  - type: textarea
+    id: note
+    attributes:
+      label: What's on your mind?
+      description: Anything is fine. We’ll help categorize it.
+      placeholder: |
+        Write anything here — bug, idea, question, feedback.
+    validations:
+      required: true
