### 2025-10-22 3f95701
Replaced signal filtering with data flow in doc.
