+++ docs/features.rst
@@ -99,7 +99,7 @@ to another.
 +================================================+================================================+
 | *Set LED to ON when button is pressed:*                                                         |
 +------------------------------------------------+------------------------------------------------+
-| ``digitalWrite(12, digitalRead(2)/4);``        | ``button >> led;``                             |
+| ``digitalWrite(12, digitalRead(2));``          | ``button >> led;``                             |
 +------------------------------------------------+------------------------------------------------+
 | *Set LED to ON when input sensor is high:*                                                      |
 +------------------------------------------------+------------------------------------------------+
