### 2025-10-22 aac8750
Fixed typo in doc, as reported in #152.
