### 2025-10-22 e095869
Added first draft of community guidelines on documentation website.
