+++ docs/index.rst
@@ -10,7 +10,7 @@ Essentials
    why_plaquette
    features
    reference
-   PDF manual <https://raw.githubusercontent.com/SofaPirate/Plaquette/master/extras/Plaquette-Manual.pdf>
+   community
 
 Guide
 =====
@@ -61,6 +61,7 @@ Related Info
 
    Video tutorials <https://www.youtube.com/playlist?list=PLO0YogNIPwXwEsNsoQSKeCdYlepWFrYip>
    GitHub repository <https://github.com/SofaPirate/Plaquette/>
+   PDF manual <https://raw.githubusercontent.com/SofaPirate/Plaquette/master/extras/Plaquette-Manual.pdf>
    credits
    license
    Doxygen Documentation <http://plaquette.org/Plaquette/doxygen/>
