### 2025-10-22 0c8e0b0
Adjustment to titles of issue report templates.
