+++ .github/ISSUE_TEMPLATE/feature_request.yml
@@ -1,4 +1,4 @@
-name: "💡 Idea / feature"
+name: "💡 Feature idea"
 description: Suggest an improvement or new capability
 title: "[Feature] <short summary>"
 labels: ["enhancement", "triage"]
