### 2025-10-22 313590b
Simplification of contribution instructions.
