### 2025-10-22 256d293
Fix in issues types in community guidelines.
