### 2025-10-22 bd9930e
Reviewed / simplified issues templates.
