+++ .github/ISSUE_TEMPLATE/quick_note.yml
@@ -1,7 +1,7 @@
 name: "✏️ Quick note"
-description: Leave a quick thought, glitch, idea, or question — one box only
+description: Leave a quick thought, glitch, idea, or question
 title: "[Note] <short summary>"
-labels: ["triage"]
+labels: []
 assignees: []
 body:
   - type: textarea
@@ -10,6 +10,6 @@ body:
       label: What's on your mind?
       description: Anything is fine. We’ll help categorize it.
       placeholder: |
-        Write anything here — bug, idea, question, feedback.
+        Write anything here (bug, idea, question, feedback).
     validations:
       required: true
