### 2025-10-22 95d99f9
Fixed table in CONTRIBUTING.md.
