### 2025-10-22 dc990a4
Adjusted community guidelines in documentation.
