+++ docs/community.rst
@@ -5,16 +5,14 @@ Community Guidelines
 
 Plaquette is an open-source software used by artists, makers, researchers, and developers interested
 in creative computation. The goal of these guidelines is to help everyone collaborate and
-communicate in a friendly and inclusive way — whether you are reporting a bug, suggesting a
+communicate in a friendly and inclusive way, whether you are reporting a bug, suggesting a
 feature, or sharing your ideas.
 
 If you participate in any Plaquette space (issues, pull requests, or discussions on GitHub),
 please take a moment to read and follow these principles.
 
-----
-
 Be Respectful and Inclusive
-===========================
+---------------------------
 
 Everyone is welcome, regardless of experience level, background, or identity.
 We ask that you be kind, patient, and supportive of others’ learning processes.
@@ -25,10 +23,8 @@ All interactions on GitHub are covered by the
 If you ever experience or witness behavior that makes you uncomfortable, please reach out to the
 maintainers directly.
 
-----
-
 Contributing
-============
+------------
 
 Plaquette grows through its community of users and contributors.
 You can help by:
@@ -41,28 +37,24 @@ You can help by:
 All contribution details and coding conventions are described in
 `CONTRIBUTING.md <https://github.com/SofaPirate/Plaquette/blob/main/CONTRIBUTING.md>`_.
 
-----
-
 Reporting Bugs and Requesting Features
-======================================
+--------------------------------------
 
 To make it easier for everyone:
 
 * First, check the existing issues on GitHub to see if your topic already exists.
 * If not, open a new issue and choose one of these forms:
 
-  - **🐞 Bug report** — something isn’t working.
-  - **💡 Feature idea** — you’d like to suggest an improvement.
-  - **🙋 Help needed** — you’re stuck and need a hand.
-  - **✏️ Quick note** — a short comment, idea, or thought.
+  - **Bug report** — something isn’t working.
+  - **Feature idea** — you’d like to suggest an improvement.
+  - **Help needed** — you’re stuck and need a hand.
+  - **Quick note** — a short comment, idea, or thought.
 
 All discussions happen transparently on the
 `issues tracker <https://github.com/SofaPirate/Plaquette/issues>`_.
 
-----
-
 Getting Support
-===============
+---------------
 
 If you need help using Plaquette:
 
@@ -72,18 +64,8 @@ If you need help using Plaquette:
 
 There’s no wrong question!
 
-----
-
 License and Attribution
-=======================
+-----------------------
 
 Plaquette is free software under the **GNU GPL v3 or later** license.
 See the `LICENSE <https://github.com/SofaPirate/Plaquette/blob/main/LICENSE>`_ file for details.
-
-----
-
-Thank You ❤️
-============
-
-Your curiosity, ideas, and feedback make Plaquette possible.
-Every question, suggestion, or bug report helps shape a better library for creative coding.
