### 2025-10-22 fea85dc
Minor change in list of issues types.
