### 2025-10-22 61a26b7
Fixed documentation of wrap() function following #152.
