+++ _sources/community.rst.txt
@@ -1,0 +1,71 @@
+.. _community:
+
+Community Guidelines
+====================
+
+Plaquette is an open-source software used by artists, makers, researchers, and developers interested
+in creative computation. The goal of these guidelines is to help everyone collaborate and
+communicate in a friendly and inclusive way, whether you are reporting a bug, suggesting a
+feature, or sharing your ideas.
+
+If you participate in any Plaquette space (issues, pull requests, or discussions on GitHub),
+please take a moment to read and follow these principles.
+
+Be Respectful and Inclusive
+---------------------------
+
+Everyone is welcome, regardless of experience level, background, or identity.
+We ask that you be kind, patient, and supportive of others’ learning processes.
+
+All interactions on GitHub are covered by the
+`Code of Conduct <https://github.com/SofaPirate/Plaquette/blob/master/CODE_OF_CONDUCT.md>`_.
+
+If you ever experience or witness behavior that makes you uncomfortable, please reach out to the
+maintainers directly.
+
+Contributing
+------------
+
+Plaquette grows through its community of users and contributors.
+You can help by:
+
+* Reporting bugs or unexpected behavior.
+* Suggesting new features or improvements.
+* Fixing typos, improving examples, or clarifying documentation.
+* Sharing creative projects that use Plaquette.
+
+All contribution instructions and coding conventions are described in our
+`Contribution Guidelines <https://github.com/SofaPirate/Plaquette/blob/master/CONTRIBUTING.md>`_.
+
+Reporting Bugs and Requesting Features
+--------------------------------------
+
+To make it easier for everyone:
+
+* First, check the existing issues on GitHub to see if your topic already exists.
+* If not, open a new issue and choose one of these forms:
+
+  - **Bug report** if something isn’t working.
+  - **Feature idea** if you’d like to suggest an improvement.
+  - **Help needed** if you’re stuck and need a hand.
+  - **Quick note** for a short comment, idea, or thought.
+
+All discussions happen transparently on the
+`issues tracker <https://github.com/SofaPirate/Plaquette/issues>`_.
+
+Getting Support
+---------------
+
+If you need help using Plaquette:
+
+* Start by browsing the documentation at https://plaquette.org/
+* If you still have trouble, open a **Help needed** issue on GitHub and describe what you’re trying to
+  do.
+
+There’s no wrong question!
+
+License and Attribution
+-----------------------
+
+Plaquette is free software under the **GNU GPL v3 or later** license.
+See the `LICENSE <https://github.com/SofaPirate/Plaquette/blob/master/LICENSE.md>`_ file for details.
