+++ _sources/wrap.rst.txt
@@ -16,9 +16,9 @@ Restricts a value to an interval [low, high) by wrapping it around.
 +----------------------------+-----------+
 | ``wrap(5.0, 1.0, 5.0)``    | ``1.0``   |
 +----------------------------+-----------+
-| ``wrap(10.0, 1.0, 5.0)``   | ``1.0``   |
+| ``wrap(6.0, 1.0, 5.0)``    | ``2.0``   |
 +----------------------------+-----------+
-| ``wrap(13.0, 1.0, 5.0)``   | ``3.0``   |
+| ``wrap(-1.0, 1.0, 5.0)``   | ``3.0``   |
 +----------------------------+-----------+
 
 Two alternative versions are provided:
