### 2025-10-22 f81dd30
Sync with latest fixes.
