### 2025-10-22 SofaPirate/Plaquette#152 @sofian
I added community guidelines:
- In documentation: https://plaquette.org/community.html
- In the CONTRIBUTING.md file https://github.com/SofaPirate/Plaquette/blob/master/CONTRIBUTING.md
- Added a code of conduct: https://github.com/SofaPirate/Plaquette/blob/master/CODE_OF_CONDUCT.md
- Added templates for issues (bugs, features, questions, notes)

I also fixed typos in documentation and fixed wrap() function documentation.
