### 2025-10-22 550c1cb
Minor adjustments to community guidelines.
