### 2025-10-22 8c19d67
Re-sync with latest fixes.
