### 2025-10-24 eb9f1a9
Added comparison in paper with existing Arduino libraries and Arduino timers (cf. #152).
