+++ extras/joss-paper/paper.md
@@ -42,7 +42,11 @@ Beyond its technical contributions, *Plaquette* serves as a bridge between scien
 
 In contrast, dataflow software popular within scientific and creative communities working with interactive media such as Pure Data, Max, and TouchDesigner provide powerful models for composing with signals, but are too memory and CPU intensive to run on constrained hardware. Similarly, scientific tools such as Python’s NumPy/SciPy, Matlab, or R, while offering rich signal analysis tools, are not designed for real-time processing on embedded devices.
 
-*Plaquette* addresses these gaps by bringing the expressive power of dataflow signal-based programming into a lightweight object-oriented framework optimized for microcontrollers (i.e., requiring minimal CPU and memory usage). It enables intuitive handling of signals, providing efficient implementations of core data processing functions such as peak detection, normalization, scaling, and smoothing (low-pass). This enables researchers in art and science to focus on experimentation and expressivity while ensuring accurate and reliable real-time performance on resource-constrained platforms.
+Several Arduino libraries specialize in particular aspects of real-time signal processing such as filtering (e.g., [arduinoFFT](https://docs.arduino.cc/libraries/arduinofft/), [DataTome](https://docs.arduino.cc/libraries/datatome/), [FIR-filter](https://docs.arduino.cc/libraries/fir-filter/)), generating waveforms (e.g., [FunctionGenerator](https://docs.arduino.cc/libraries/functiongenerator/), [SyncWaveformsLib](https://docs.arduino.cc/libraries/syncwaveformslib/)), or timing (e.g., [arduino-timer](https://docs.arduino.cc/libraries/arduino-timer/), [Chrono](https://docs.arduino.cc/libraries/chrono/), [SoftTimer](https://docs.arduino.cc/libraries/softtimer/)). While these tools are effective within their domains, they address isolated functionalities and do not share a common programming model. Finally, Arduino’s hardware timers provide precise low-level control but require complex configuration of prescalers, registers, and interrupt handlers. These limitations of existing Arduino
+libraries and hardware timers hamper creative expression and experimentation.
+
+*Plaquette* addresses these gaps by bringing the expressive power of dataflow signal-based programming into a lightweight object-oriented framework optimized for microcontrollers (i.e., requiring minimal CPU and memory usage). It enables intuitive handling of signals, providing efficient implementations of core data processing functions such as peak detection, normalization, scaling, and smoothing (low-pass), all under a unified framework. This enables researchers in art and science to focus on experimentation and expressivity while ensuring accurate and reliable real-time performance on resource-constrained platforms. Additional functionalities that expand on Plaquette's core,
+such as advanced data filtering and event management, can be implemented as external Plaquette libraires.
 
 The framework provides a strong foundation for workshop-based research-creation projects, where participants often have diverse levels of technical skills. Its accessibility ensures that beginners can quickly grasp and apply core concepts, while its efficient, expressive, and extensible architecture support the needs of advanced users. This makes it particularly well-suited for collaborative prototyping in media arts, design, and human-computer interaction, where embodied and situated practices require adaptable tools.
 
