### 2025-10-31 SofaPirate/Plaquette#153 @WarmCyan
[JOSS Review] Minor typos and suggestions
Here's my review for https://github.com/openjournals/joss-reviews/issues/9144

To echo the other JOSS reviewer, this is a really cool project! I particularly like the syntax sugar that comes from overloading the `>>` operator, it's an elegant design decision and legitimately makes the code easier to read at a glance, so well done!

Installation steps, getting started, and examples were all clear, I was able to get set up and run the snippets on my arduino without any trouble.

I ran into one very minor compilation issue, I raise this mostly just for your awareness because it may have simply been a weird setup on my machine and was immediately obvious/easy to fix: running the literal example under https://plaquette.org/waves.html#types-of-waves complained about `Wave square(SQUARE, 1.0)`
```
error: 'pq::Wave square' redeclared as different kind of symbol
(...)
in file included from
.arduino15/packages/arduino/hardware/avr/1.8.6/cores/arduino/Arduino.h:26:0,
(...)
.arduino15/packages/arduino/tools/avr-gcc/7.3.0-atmel3.6.1-ardunio7/avr/include/math.h:203:15: note: previous declaration 'double square(double)' extern double square(double __x) __ATTR_CONST__;
```
So it looks like a math library already has a `square` defined. It was easy enough to just change the name in the sketch, and I'm not sure how widespread of an issue this would be (the only board I have to test on is an arduino), but possibly worth either a brief callout or slightly modifying the names. Your call if you think that's necessary!

One minor nitpick in your heartbeat example under https://plaquette.org/waves.html#wave-addition, you use `led.put(heartBeat);` instead of `heartBeat >> led;`, which I saw later on in the advanced usage docs is the OOP style API usage. I would either change it to `heartBeat >> led;` to remain consistent with all the other examples up until the later section, or at least explicitly call out that that's what's happening, just so it isn't confusing to anyone the first time they look through it.

The only other documentation comments I have is a small typo in Advanced Usage, under https://plaquette.org/advanced.html#using-plaquette-as-an-external-library
"After this step, you is then responsible" -> "After this step, you are then responsible"

Finally there's a minor typo in the paper itself on line 92 of the draft, a missing space:
"period ofoscillation" -> "period of oscillation"
