### 2025-11-02 338f3c6
Fixed typos as suggested in #153.
