### 2025-11-02 53f3002
Fixed typo in paper as reported in #153.
