### 2025-11-02 95fba6f
Sync with latest changes - fixes suggested from JOSS review.
