### 2025-11-02 SofaPirate/Plaquette#153 @sofian
Thank you for the feedback and suggested fixes. I have completed all suggested changes.
