### 2025-11-03 71c2106
minor changes in paper.md
