### 2025-11-03 #9144 @danielskatz
@sofian & @thomasfredericks - I've suggested some changes to the paper/bib in https://github.com/SofaPirate/Plaquette/pull/155 - please merge this, or let me know what you disagree with, then we can continue toward publication
