### 2025-11-05 5067ba8
More typo fixes and adjustments to text.
