### 2025-11-05 bc1b049
Minor updates to paper.
