+++ extras/joss-paper/paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 # Summary
 
-*Plaquette* is an object-oriented C++ framework for interactive media on embedded systems, supporting a wide range of platforms including AVR, ARM, SAMD, STM32, and ESP32. It provides a signal-centric architecture and a suite of modular abstractions (oscillators, filters, units, and scheduling engines) that simplify the design of time-based behaviors. Its expressive syntax allows fast prototyping with multiple sensors, actuators, and real-time processes, enabling researchers as well as creative practitioners to experiment with and design complex physical computing systems.
+*Plaquette* is an object-oriented C++ framework for interactive media on embedded systems, supporting a wide range of platforms including AVR, ARM, ESP32, SAMD, and STM32. It provides a signal-centric architecture and a suite of modular abstractions (oscillators, filters, units, and scheduling engines) that simplify the design of time-based behaviors. Its expressive syntax allows fast prototyping with multiple sensors, actuators, and real-time processes, enabling researchers as well as creative practitioners to experiment with and design complex physical computing systems.
 
 Beyond its technical contributions, *Plaquette* serves as a bridge between scientific research and creative practice. Its application to interdisciplinary projects involving affective biofeedback and robotic behaviors demonstrates how the framework’s flexible and robust infrastructure supports creativity and experimentation across interactive media.
 
@@ -40,7 +40,7 @@ Beyond its technical contributions, *Plaquette* serves as a bridge between scien
 
 *Plaquette* is designed for research—specifically, practice-based research [@Candy2018-PracticeBased] and research-creation [@Loveless2019-How]—in embedded interactive media, with applications ranging from affective computing and digital lutherie to robotic art, interactive installation, connected objects, and performance. The [Arduino](https://arduino.cc) [@Banzi2022-Getting] open-source platform for physical computing anchors a large and active ecosystem; it remains the go-to environment for artists, makers, and researchers working with embedded interactive media. Its accessibility, community, libraries, and hardware options have made it the most popular microcontroller platform worldwide. Yet, Arduino's core library is not optimized for real-time signal processing: it lacks an object-oriented design, provides limited abstractions for managing concurrent events, and often requires manipulating raw numerical values, making interaction design unintuitive and hindering expressive experimentation.
 
-In contrast, dataflow software popular within scientific and creative communities working with real-time generative or interactive media (such as Pure Data, Max, and TouchDesigner) provide powerful models for composing with signals, but are too memory- and CPU-intensive to run on constrained hardware. Similarly, scientific tools such as Python’s NumPy/SciPy, Matlab, or R, offer rich signal analysis tools, but are not designed for real-time processing on embedded devices.
+In contrast, dataflow software popular within scientific and creative communities working with real-time media (such as Pure Data, Max, and TouchDesigner) provide powerful models for composing with signals, but are too memory- and CPU-intensive to run on constrained hardware. Similarly, scientific tools such as Python’s NumPy/SciPy, Matlab, or R, offer rich signal analysis tools, but are not designed for real-time processing on embedded devices.
 
[…]
