### 2025-11-06 7963ee3
Simplification of example using callback.
