+++ docs/conf.py
@@ -55,7 +55,7 @@ project = u'Plaquette'
 copyright = u'2015-2025, Plaquette'
 author = u'Sofian Audry & Thomas O. Fredericks'
 version = '0.8'
-release = '0.8.0'
+release = '0.8.1'
 language = 'en'
 exclude_patterns = ['_build']
 pygments_style = 'sphinx'
