+++ README.md
@@ -1,4 +1,4 @@
-[![arduino-library-badge](http://www.ardu-badge.com/badge/Plaquette.svg)](http://www.ardu-badge.com/Plaquette) [![Compile Examples](https://github.com/SofaPirate/Plaquette/actions/workflows/arduino-test-compile.yml/badge.svg)](https://github.com/SofaPirate/Plaquette/actions/workflows/arduino-test-compile.yml) [![AUnit Tests](https://github.com/SofaPirate/Plaquette/actions/workflows/aunit_tests.yml/badge.svg)](https://github.com/SofaPirate/Plaquette/actions/workflows/aunit_tests.yml) [![Arduino Lint](https://github.com/SofaPirate/Plaquette/actions/workflows/arduino-lint.yml/badge.svg)](https://github.com/SofaPirate/Plaquette/actions/workflows/arduino-lint.yml)
+[![arduino-library-badge](http://www.ardu-badge.com/badge/Plaquette.svg)](http://www.ardu-badge.com/Plaquette) [![Compile Examples](https://github.com/SofaPirate/Plaquette/actions/workflows/arduino-test-compile.yml/badge.svg)](https://github.com/SofaPirate/Plaquette/actions/workflows/arduino-test-compile.yml) [![AUnit Tests](https://github.com/SofaPirate/Plaquette/actions/workflows/aunit_tests.yml/badge.svg)](https://github.com/SofaPirate/Plaquette/actions/workflows/aunit_tests.yml) [![Arduino Lint](https://github.com/SofaPirate/Plaquette/actions/workflows/arduino-lint.yml/badge.svg)](https://github.com/SofaPirate/Plaquette/actions/workflows/arduino-lint.yml) [![DOI](https://zenodo.org/badge/36823798.svg)](https://doi.org/10.5281/zenodo.17544842)
 
 Plaquette is an object-oriented, user-friendly, signal-centric programming
 framework for **creative physical computing**. It promotes **expressiveness** over
