### 2025-11-06 #9144 @sofian
@danielskatz I have completed the last steps!

Released version tag: v0.8.1
Zenodo DOI: 10.5281/zenodo.17544843
