### 2025-11-10 #9144 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
